% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catR.R
\name{catR}
\alias{catR}
\title{Cats in R console.}
\usage{
catR()
}
\value{
A photo of a cat is downloaded and displayed in the plot window.
}
\description{
Cats in R console.
}
\examples{
catR::catR()
}
\author{
\href{https://github.com/melissanjohnson/}{Melissa Innerst},
\href{https://github.com/IndrajeetPatil/}{Indrajeet Patil}
}
