% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kittypics_df_maker.R
\name{kitty_pics_df}
\alias{kitty_pics_df}
\title{Prepare a dataframe with URLs for cat pics.}
\usage{
kitty_pics_df(url, ...)
}
\arguments{
\item{url}{URL for the source from which individual image URLs are to be
extracted.}

\item{...}{Currently ignored.}
}
\description{
Prepare a dataframe with URLs for cat pics.
}
\examples{
kittyR::kitty_pics_df("https://unsplash.com/s/photos/kitten/")
}
