% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kittyR-package.R
\docType{package}
\name{kittyR-package}
\alias{kittyR}
\alias{kittyR-package}
\title{kittyR: Kitty Pictures and Meows from R Console}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Get pictures of cats and their meows from your R
    console.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/IndrajeetPatil/kittyR}
  \item Report bugs at \url{https://github.com/IndrajeetPatil/kittyR/issues}
}

}
\author{
\strong{Maintainer}: Indrajeet Patil \email{patilindrajeet.science@gmail.com} (\href{https://orcid.org/0000-0003-1995-6531}{ORCID}) (@patilindrajeets) [copyright holder]

}
\keyword{internal}
