% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kittyR.R
\name{kittyR}
\alias{kittyR}
\title{Cats in R console.}
\usage{
kittyR(meow = TRUE, sound = 3)
}
\arguments{
\item{meow}{Logical that decides whether to play a meow sound along with the
picture of a cat.}

\item{sound}{A character string or a number specifying what sound to be
played by either specifying one of the built in sounds, specifying the path
to a \code{.wav} file or specifying an url. There are currently 6 meows
included. The default is \code{1}.}
}
\value{
A photo of a cat is downloaded and displayed in the plot window.
}
\description{
Cats in R console.
}
\examples{
kittyR::kittyR(meow = FALSE)
}
\author{
\href{https://github.com/IndrajeetPatil/}{Indrajeet Patil}
}
