% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meowR.R
\name{meowR}
\alias{meowR}
\title{Play a short kitty meow sound}
\usage{
meowR(sound = 3)
}
\arguments{
\item{sound}{A character string or a number specifying which sound is to be
played by specifying one of the built-in sounds (stored in \code{.wav} files).
There are currently 6 meows included. The default is third one (\code{3}).}
}
\value{
No return value, called for side effect (sound).
}
\description{
\code{meowR} is inspired by \code{beepr::beep}, which plays a short sound to notify
you, for example, when a script has finished. Instead of a beep, this
function produced a meow.
}
\examples{
\donttest{
kittyR::meowR(sound = 4)
}
}
