% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kittypics_df_maker.R
\name{kitty_pics_df}
\alias{kitty_pics_df}
\title{Prepare a dataframe with URLs for cat pics.}
\usage{
kitty_pics_df(url, ...)
}
\arguments{
\item{url}{URL for the source from which individual image URLs are to be
extracted.}

\item{...}{Currently ignored.}
}
\value{
A tibble containing extracted details from the provided URL.
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("maturing")}

Mostly useful for internal workings of this package, this function extracts
images at given URL and returns it in a dataframe.
}
\examples{
set.seed(123)
kittyR::kitty_pics_df("https://unsplash.com/s/photos/kitten/")
}
