% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meowR.R
\name{meowR}
\alias{meowR}
\title{Play a short kitty meow sound}
\usage{
meowR(sound = 3)
}
\arguments{
\item{sound}{A character string or a number specifying what sound to be
played by either specifying one of the built in sounds, specifying the path
to a \code{.wav} file or specifying an URL. There are currently 6 meows
included. The default is \code{3}.}
}
\value{
No return value, called for side effect (sound).
}
\description{
\code{meowR} is inspired by \code{beepr::beep}, which plays a short sound to notify
you, for example, when a script has finished. Instead of a beep, this
function produced a meow.
}
\examples{
\donttest{
# play Eno's meow
kittyR::meowR(sound = 4)
}
}
\author{
\href{https://github.com/IndrajeetPatil/}{Indrajeet Patil}
}
