% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meowR.R
\name{meowR}
\alias{meowR}
\title{Play a short kitty meow sound.}
\usage{
meowR(sound = 1)
}
\arguments{
\item{sound}{A character string or a number specifying what sound to be
played by either specifying one of the built in sounds, specifying the path
to a \code{.wav} file or specifying an url. There are currently 6 meows
included. The default is \code{1}.}
}
\description{
Play a short kitty meow sound.
}
\examples{
\dontrun{
# play Eno's meow
kittyR::meowR(sound = 3)
}
}
\author{
\href{https://github.com/IndrajeetPatil/}{Indrajeet Patil}
}
